split bindings and implementation, unsafe
