+++ tests/compiler/wasi/abort.ts
@@ -1,4 +1,4 @@
-import "bindings/wasi";
+import "wasi";
 
 export function test(): void {
   assert(false, "the message");
