+++ tests/compiler/wasi/seed.ts
@@ -1,4 +1,4 @@
-import "bindings/wasi";
+import "wasi";
 
 export function test(): f64 {
   return Math.random();
